# openapi-spec.json
{
  "openapi": "3.1.1",
  "paths": {
...
    "/api/v1/app/list": {
      "get": {
        "operationId": "OpenAppController_getAppList",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "description": "page number",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page_size",
            "required": true,
            "in": "query",
            "description": "page size",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppListDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get app list",
        "tags": [
          "IP Portal"
        ]
      }
    },
    "/api/v1/app/info": {
      "get": {
        "operationId": "OpenAppController_getAppInfo",
        "parameters": [
          {
            "name": "app-id",
            "required": true,
            "in": "header",
            "description": "App ID, please contact us to get the app id",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "required": false,
            "in": "header",
            "description": "JWT Authorization, if set,this field is using to verify if user is admin of the app",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "App info",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppInfoDto"
                }
              }
            }
          }
        },
        "summary": "Get app info",
        "tags": [
          "IP Portal"
        ]
      }
    },
...
    "/api/v1/app/create": {
      "post": {
        "operationId": "OpenAppController_createApp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAppDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create app",
        "tags": [
          "IP Portal"
        ]
      }
    },
    "/api/v1/app/preview": {
      "post": {
        "operationId": "OpenAppController_previewApp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAppDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppInfoDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Preview app",
        "tags": [
          "IP Portal"
        ]
      }
    },
    "/api/v1/app/delete": {
      "post": {
        "operationId": "OpenAppController_deleteApp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteAppDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete app",
        "tags": [
          "IP Portal"
        ]
      }
    },
    "/api/v1/app/update-app": {
      "post": {
        "operationId": "OpenAppController_updateApp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAppDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update app",
        "tags": [
          "IP Portal"
        ]
      }
    },
    "/api/v1/app/top-ip-list": {
      "get": {
        "operationId": "OpenAppController_getTopIpList",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TopIpSummaryDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get top IP list",
        "tags": [
          "IP Portal"
        ]
      }
    },
    "/api/v1/app/request-creator": {
      "post": {
        "description": "Sends an application to become a creator on the platform with contact details and project ideas",
        "operationId": "OpenAppController_requestCreator",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestCreatorDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Request sent successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestCreatorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Request to become a creator on the platform",
        "tags": [
          "IP Portal"
        ]
      }
    },
    "/api/v1/app/approve-creator": {
      "post": {
        "description": "Approves a creator application, enables IP creation permissions, and sends a confirmation email",
        "operationId": "OpenAppController_approveCreator",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveCreatorDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Creator approved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApproveCreatorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Approve a creator application",
        "tags": [
          "IP Portal"
        ]
      }
    },
    "/api/v1/app/lookup-by-subdomain/{subdomain}": {
      "get": {
        "operationId": "OpenAppController_lookupBySubdomain",
        "parameters": [
          {
            "name": "subdomain",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "App info",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppInfoDto"
                }
              }
            }
          }
        },
        "summary": "Lookup app by subdomain",
        "tags": [
          "IP Portal"
        ]
      }
    },
    "/api/v1/app/info/{appId}": {
      "get": {
        "description": "Get app info by app id",
        "operationId": "OpenAppController_getAppInfoByAppId",
        "parameters": [
          {
            "name": "appId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "required": false,
            "in": "header",
            "description": "JWT Auth orization, if set,this field is using to verify if user is admin of the app",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "App info",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppInfoDto"
                }
              }
            }
          }
        },
        "summary": "Get app info by app id",
        "tags": [
          "IP Portal"
        ]
      }
    },
...
  },
  "info": {
    "title": "3 Body Labs API Reference",
    "description": "This is the API reference for 3 Body Labs",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "in": "header",
        "name": "Authorization",
        "description": "JWT Authorization"
      }
    },
    "schemas": {
      "SubscriptionPlanDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "The name of the plan",
            "enum": [
              "none",
              "Free",
              "Starter",
              "Basic",
              "Premium",
              "Custom"
            ]
          },
          "period": {
            "type": "string",
            "description": "The period of the plan",
            "enum": [
              "monthly",
              "yearly"
            ]
          },
          "price": {
            "type": "number",
            "description": "The price of the plan"
          },
          "price_per_credit": {
            "type": "number",
            "description": "The price per credit of the plan, using to purchase external credits"
          },
          "price_id": {
            "type": "string",
            "description": "Stripe price id of the plan"
          },
          "credit_per_month": {
            "type": "number",
            "description": "The credits per month of the plan"
          },
          "video_convert_max_seconds": {
            "type": "number",
            "description": "The maximum seconds of animation convert of the plan"
          },
          "credit_consume_every_second": {
            "type": "number",
            "description": "The credits consume every second of the plan"
          },
          "face_swap_consume_every_second": {
            "type": "number",
            "description": "The credits consume every second of the plan"
          },
          "generate_video_consume_every_second": {
            "type": "number",
            "description": "The credits consume every second of the plan"
          },
          "generate_image_consume_per_image": {
            "type": "number",
            "description": "The credits consume per image of the plan"
          },
          "extra_info": {
            "type": "object",
            "description": "The rights of the plan",
            "required": [],
            "properties": {
              "max_animate_seconds": {
                "description": "The maximum seconds of animation convert of the plan",
                "type": "number"
              },
              "max_generate_videos": {
                "description": "The maximum seconds of generate videos of the plan",
                "type": "number"
              },
              "max_generate_images": {
                "description": "The maximum seconds of generate images of the plan",
                "type": "number"
              }
            }
          }
        },
        "required": [
          "name",
          "period",
          "price",
          "price_per_credit",
          "price_id",
          "credit_per_month",
          "video_convert_max_seconds",
          "credit_consume_every_second",
          "face_swap_consume_every_second",
          "generate_video_consume_every_second",
          "generate_image_consume_per_image"
        ]
      },
      "UserInfoDTO": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "usernameShorted": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          },
          "credit": {
            "type": "number"
          },
          "emailConfirmed": {
            "type": "boolean"
          },
          "avatar": {
            "type": "string"
          },
          "invited_by": {
            "type": "string"
          },
          "agent_user": {
            "type": "string"
          },
          "subscription_info": {
            "$ref": "#/components/schemas/SubscriptionPlanDto"
          },
          "giggle_wallet_address": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "followers": {
            "type": "number"
          },
          "following": {
            "type": "number"
          },
          "app_id": {
            "type": "string"
          },
          "can_create_ip": {
            "type": "boolean"
          },
          "permissions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "widget_info": {
            "type": "object"
          }
        },
        "required": [
          "username",
          "password",
          "usernameShorted",
          "email",
          "refreshToken",
          "credit",
          "emailConfirmed",
          "avatar",
          "invited_by",
          "agent_user",
          "subscription_info",
          "giggle_wallet_address",
          "description",
          "followers",
          "following",
          "app_id",
          "can_create_ip",
          "permissions",
          "widget_info"
        ]
      },
...
      "OnChainDetailDto": {
        "type": "object",
        "properties": {
          "ipAddr": {
            "type": "string",
            "description": "ip address of the ip library"
          },
          "signature": {
            "type": "string",
            "description": "signature of the ip library"
          }
        },
        "required": [
          "ipAddr",
          "signature"
        ]
      },
...
      "IpLibraryDetailDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "id of the ip library"
          },
          "name": {
            "type": "string",
            "description": "name of the ip library"
          },
          "ticker": {
            "type": "string",
            "description": "ticker of the ip library"
          },
          "description": {
            "type": "string",
            "description": "description of the ip library"
          },
          "likes": {
            "type": "number",
            "description": "likes of the ip library"
          },
          "comments": {
            "type": "number",
            "description": "comments of the ip library"
          },
          "is_user_liked": {
            "type": "boolean",
            "description": "is liked by the user"
          },
          "cover_asset_id": {
            "type": "number",
            "description": "cover asset id of the ip library"
          },
          "cover_image": {
            "type": "string",
            "description": "cover images of the ip library"
          },
          "cover_hash": {
            "type": "string",
            "description": "cover ipfs hash of the ip library"
          },
          "on_chain_detail": {
            "description": "on chain detail",
            "allOf": [
              {
                "$ref": "#/components/schemas/OnChainDetailDto"
              }
            ]
          },
          "can_purchase": {
            "type": "boolean",
            "description": "can purchase"
          },
          "creation_guide_lines": {
            "type": "string",
            "description": "creation guide lines"
          },
          "governance_right": {
            "type": "boolean",
            "description": "is governance right"
          },
          "creator_id": {
            "type": "string",
            "description": "creator id of the ip library"
          },
          "creator": {
            "type": "string",
            "description": "creator of the ip library"
          },
          "creator_description": {
            "type": "string",
            "description": "creator description of the ip library"
          },
          "creator_avatar": {
            "type": "string",
            "description": "creator avatar of the ip library"
          },
          "is_public": {
            "type": "boolean",
            "description": "is public"
          },
          "is_top": {
            "type": "boolean",
            "description": "is top"
          },
          "creator_followers": {
            "type": "number",
            "description": "creator followers of the ip library"
          },
          "token_info": {
            "description": "token info",
            "allOf": [
              {
                "$ref": "#/components/schemas/CreateIpTokenGiggleResponseDto"
              }
            ]
          },
          "ip_signature_clips": {
            "description": "signature clips of the ip library",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IpSignatureClipDto"
            }
          },
          "authorization_settings": {
            "description": "authorization settings of the ip library",
            "allOf": [
              {
                "$ref": "#/components/schemas/AuthorizationSettingsDto"
              }
            ]
          },
          "genre": {
            "description": "genre of the ip library",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GenreDto"
            }
          },
          "parent_ip_info": {
            "description": "parent ip library",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IpSummaryDto"
            }
          },
          "child_ip_info": {
            "description": "child ip library, only return 100 items",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IpSummaryDto"
            }
          },
          "on_chain_status": {
            "type": "string",
            "description": "on chain status"
          },
          "extra_info": {
            "type": "object",
            "description": "extra info"
          }
        },
        "required": [
          "id",
          "name",
          "ticker",
          "description",
          "likes",
          "comments",
          "is_user_liked",
          "cover_asset_id",
          "cover_image",
          "cover_hash",
          "on_chain_detail",
          "can_purchase",
          "creation_guide_lines",
          "governance_right",
          "creator_id",
          "creator",
          "creator_description",
          "creator_avatar",
          "is_public",
          "is_top",
          "creator_followers",
          "token_info",
          "ip_signature_clips",
          "authorization_settings",
          "genre",
          "parent_ip_info",
          "child_ip_info",
          "on_chain_status",
          "extra_info"
        ]
      },
...
      "AppListDto": {
        "type": "object",
        "properties": {}
      },
      "AppInfoDto": {
        "type": "object",
        "properties": {
          "app_id": {
            "type": "string",
            "description": "App ID"
          },
          "app_name": {
            "type": "string",
            "description": "App Name"
          },
          "description": {
            "type": "string",
            "description": "App Description"
          },
          "user_info": {
            "description": "User Info",
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/UserInfoDTO"
              }
            ]
          },
          "ip_info": {
            "description": "IP Info",
            "allOf": [
              {
                "$ref": "#/components/schemas/IpLibraryDetailDto"
              }
            ]
          },
          "is_admin": {
            "type": "boolean",
            "description": "Is Admin"
          },
          "usdc_mint": {
            "type": "string",
            "description": "USDC Mint"
          },
          "sub_domain": {
            "type": "string",
            "description": "Sub Domain"
          },
          "style_name": {
            "type": "string",
            "description": "Style Name"
          },
          "radius": {
            "type": "string",
            "description": "Radius"
          },
          "configs": {
            "type": "object",
            "description": "App configs, you can pass any configs you want"
          },
          "kline_url": {
            "type": "string",
            "description": "Kline URL"
          },
          "widgets": {
            "description": "Widgets",
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        },
        "required": [
          "app_id",
          "user_info",
          "ip_info",
          "is_admin",
          "usdc_mint",
          "sub_domain",
          "style_name",
          "radius",
          "kline_url",
          "widgets"
        ]
      },
...
      "CreateAppDto": {
        "type": "object",
        "properties": {
          "sub_domain": {
            "type": "string",
            "description": "Sub Domain"
          },
          "style_name": {
            "type": "string",
            "description": "Style Name"
          },
          "radius": {
            "type": "string",
            "description": "Radius"
          },
          "configs": {
            "type": "object",
            "description": "App configs, you can pass any configs you want"
          },
          "widgets": {
            "description": "Widgets",
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "ip_id": {
            "type": "number",
            "description": "IP ID, this ip must be shared to giggle and have no parent ip and ip owner is the same as the app owner"
          }
        },
        "required": [
          "sub_domain",
          "style_name",
          "radius",
          "widgets",
          "ip_id"
        ]
      },
      "DeleteAppDto": {
        "type": "object",
        "properties": {
          "app_id": {
            "type": "string",
            "description": "App ID"
          }
        },
        "required": [
          "app_id"
        ]
      },
      "UpdateAppDto": {
        "type": "object",
        "properties": {
          "sub_domain": {
            "type": "string",
            "description": "Sub Domain"
          },
          "style_name": {
            "type": "string",
            "description": "Style Name"
          },
          "radius": {
            "type": "string",
            "description": "Radius"
          },
          "configs": {
            "type": "object",
            "description": "App configs, you can pass any configs you want"
          },
          "widgets": {
            "description": "Widgets",
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "app_id": {
            "type": "string",
            "description": "App ID"
          },
          "ip_id": {
            "type": "number",
            "description": "IP ID"
          }
        },
        "required": [
          "sub_domain",
          "style_name",
          "radius",
          "widgets",
          "app_id",
          "ip_id"
        ]
      },
      "TopIpSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "id of the ip library"
          },
          "name": {
            "type": "string",
            "description": "name of the ip library"
          },
          "ticker": {
            "type": "string",
            "description": "ticker of the ip library"
          }
        },
        "required": [
          "id",
          "name",
          "ticker"
        ]
      },
      "RequestCreatorDto": {
        "type": "object",
        "properties": {
          "full_name": {
            "type": "string",
            "description": "Full name of the requester",
            "example": "John Doe"
          },
          "email": {
            "type": "string",
            "description": "Email of the requester",
            "example": "johndoe@example.com"
          },
          "company": {
            "type": "string",
            "description": "Company name",
            "example": "Example Inc."
          },
          "website": {
            "type": "string",
            "description": "Website URL",
            "example": "https://example.com"
          },
          "social_media": {
            "type": "string",
            "description": "Social media links (comma separated)",
            "example": "https://twitter.com/example, https://linkedin.com/in/example"
          },
          "description": {
            "type": "string",
            "description": "Description of the request or project ideas",
            "example": "I would like to create apps related to digital art and NFTs..."
          }
        },
        "required": [
          "full_name",
          "email",
          "description"
        ]
      },
      "RequestCreatorResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Whether the request was successful",
            "example": true
          },
          "message": {
            "type": "string",
            "description": "Message to display to the user",
            "example": "Your request has been sent to the creator"
          }
        },
        "required": [
          "success",
          "message"
        ]
      },
      "ApproveCreatorDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Email of the creator to approve",
            "example": "johndoe@example.com"
          }
        },
        "required": [
          "email"
        ]
      },
      "ApproveCreatorResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Whether the request was successful",
            "example": true
          },
          "message": {
            "type": "string",
            "description": "Message to display to the user",
            "example": "Creator has been approved successfully"
          }
        },
        "required": [
          "success",
          "message"
        ]
      },
...
    }
  },
...
}